Refactor DNS and TLS inspection modules
