+++ AGENTS.md
@@ -116,6 +116,7 @@ metadata/update/DNS/TLS inspection modes, and executes requests via `src/http`.
 - Rust TLS version options accept only TLS 1.2 and TLS 1.3; legacy TLS 1.0/1.1 values are rejected consistently for CLI flags, config, WebSocket, and inspection paths.
 - Rustls is built with `aws-lc-rs` and `prefer-post-quantum`; keep post-quantum key exchange preference enabled unless there is a concrete compatibility or provider reason to disable it.
 - TLS inspection keeps its custom verifier for certificate display and OCSP capture, but Rustls protocol-version selection and PEM/client-auth material parsing should stay centralized in `src/tls/mod.rs`.
+- TLS inspection orchestration stays in `src/tls/inspect.rs`; DER reading, certificate/OCSP parsing, and display-chain shaping live under `src/tls/inspect/cert.rs` and `src/tls/inspect/der.rs`, while TLS inspection output rendering lives in `src/tls/inspect/render.rs`.
 - Direct HTTPS requests discover HTTP/3 via DNS HTTPS/SVCB records advertising `h3`, race QUIC setup against the normal TCP/TLS `h2`/`http/1.1` ALPN path, and send the request once on the winning transport. With `--dns-server`, HTTPS-record discovery uses that custom UDP or DoH resolver; without it, discovery uses the platform resolver, matching normal A/AAAA lookup behavior. Proxy, Unix socket, explicit HTTP version, and gRPC/WebSocket paths bypass auto-H3. Explicit `--http 1`, `--http 2`, and `--http 3` remain forced protocol selections, not version caps. Auto-H3 HTTPS/SVCB discovery is opportunistic and must not delay normal connection setup: start it in parallel with normal A/AAAA lookup and TCP/TLS setup, race QUIC only when a usable `h3` candidate is discovered before TCP/TLS wins, and skip SVCB target-name resolution that would require an extra DNS wait unless address hints, an origin target, or an IP target are already available. The auto-H3 QUIC/TCP race shares one started `--connect-timeout` budget across both branches; do not create fresh branch-local connect budgets after the fallback delay.
 - Automatic HTTP/3 stores learned alternatives in a bounded sharded JSON cache under the user cache directory (`http3/<prefix>/<hash>.json`). Cache entries are scoped by normalized origin and resolver key, store alternative authorities instead of IP addresses, expire by HTTPS/SVCB TTL or `Alt-Svc` `ma` with a hard retention cap, and should not be used for proxy, Unix socket, IP-literal, non-HTTPS, explicit `--http`, or gRPC/WebSocket paths. Fresh HTTPS/SVCB candidates take precedence when available promptly, but cached alternatives race while slower HTTPS/SVCB discovery continues so learned `Alt-Svc` entries are not hidden behind unsupported or slow HTTPS-record lookups. `Alt-Svc` learning is disabled for `--insecure` responses.
 - The custom transport should preserve reqwest's safe default retries for HTTP/2/3 protocol NACKs such as remote `REFUSED_STREAM`, remote `GOAWAY(NO_ERROR)`, and HTTP/3 connection timeout signals, but only when the request body can be replayed.
@@ -153,6 +154,7 @@ metadata/update/DNS/TLS inspection modes, and executes requests via `src/http`.
 - DoH responses are capped at 1 MiB while buffering; oversized responses must fail with a DNS error instead of growing memory without bound.
 - DoH resolver URLs are queried with RFC 8484 `application/dns-message` POST requests first, falling back to Google-style JSON DoH for compatibility; keep the same response size cap on both paths.
 - DNS implementation details are centralized under `src/dns`: `custom.rs` owns `--dns-server` dispatch, `wire.rs` owns DNS packet encoding/parsing, and `doh.rs` owns DoH querying. Request, TLS, WebSocket, and inspection code should reuse those helpers instead of branching on DoH vs UDP locally.
+- DNS inspection lookup orchestration stays in `src/dns/inspect.rs`; RDATA formatting and DoH generic-RDATA normalization live in `src/dns/inspect/rdata.rs`, while DNS inspection output rendering lives in `src/dns/inspect/render.rs`.
 - HTTP request, DoH HTTPS, and WebSocket `wss://` connections use rustls' platform verifier to match the old reqwest path and avoid loading native roots into a Rust store during request/DNS timing; request-path TLS config still needs to preserve `--ca-cert`, mTLS, `--insecure`, and TLS min/max behavior.
 - GitHub Actions run Cargo fmt, clippy, unit tests, and the Rust integration suite. Release builds Cargo archives named for the self-updater, Linux GNU binaries are built with a prebuilt `cargo-zigbuild` against an explicit glibc 2.28 floor, Windows release binaries use the static MSVC CRT, and each archive is uploaded with a SHA-256 sidecar. The release workflow supports manual dry runs via `workflow_dispatch`, uploading archives as workflow artifacts unless explicitly told to upload to an existing GitHub Release. Release builds set `FETCH_VERSION` from the release tag/manual version so the compiled binary reports the published or test version; `Cargo.toml` intentionally remains `0.0.0` unless crate publishing becomes a goal. Local builds derive `FETCH_VERSION` from a matching `v*` git tag, then `git describe`, then `v0.0.0-dev`. Build info marks `vcs.modified` from tracked Git changes only, ignoring untracked files such as local Zig caches.
 - Rust is pinned to 1.96.0 via `rust-toolchain.toml`; keep `Cargo.toml`'s `rust-version` and GitHub Actions toolchain setup aligned with that version.
